% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_toml.R
\name{from_toml}
\alias{from_toml}
\title{Convert \code{Toml} to an R object}
\usage{
from_toml(x)
}
\arguments{
\item{x}{a \code{Toml} object.}
}
\value{
a list
}
\description{
Use \code{from_toml()} to convert a \code{Toml} document to an R object.
Note that that due to the encoding of values in the TOML specification
a perfect round trip from TOML to R is not always possible.
}
\examples{
from_toml(toml(hello = "world"))
}
