% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomledit-package.R
\docType{package}
\name{tomledit-package}
\alias{tomledit}
\alias{tomledit-package}
\title{tomledit: Parse, Read, and Edit 'TOML'}
\description{
A toolkit for working with 'TOML' files in R while preserving formatting, comments, and structure. 'tomledit' enables serialization of R objects such as lists, data.frames, numeric, logical, and date vectors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://extendr.rs/tomledit/}
  \item \url{https://github.com/extendr/tomledit}
  \item Report bugs at \url{https://github.com/extendr/tomledit/issues}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

Other contributors:
\itemize{
  \item Devin Pastoor \email{devin.pastoor@gmail.com} [contributor]
}

}
\keyword{internal}
